move better_torch_load to torch_utils
