+++ yolov5/__init__.py
@@ -1,4 +1,4 @@
-from yolov5.helpers import YOLOv5, better_torch_load
+from yolov5.helpers import YOLOv5
 from yolov5.helpers import load_model as load
 
 __version__ = "5.0.4"
