+++ yolov5/helpers.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from yolov5.models.yolo import Model
 from yolov5.utils.general import set_logging
 from yolov5.utils.google_utils import attempt_download
-from yolov5.utils.torch_utils import torch
+from yolov5.utils.torch_utils import better_torch_load, torch
 
 
 def load_model(model_path, device=None, autoshape=True, verbose=False):
@@ -46,52 +46,6 @@ def load_model(model_path, device=None, autoshape=True, verbose=False):
         model = model.autoshape()
 
     return model
-
-
-def better_torch_load(f, map_location=None, pickle_module=pickle, **pickle_load_args):
-    """
-    Loads an object saved with torch.save from a file.
-
-    Same interface with https://pytorch.org/docs/stable/generated/torch.load.html
-    but better yolov5 model loading handling.
-
-    Args:
-        f: a file-like object (has to implement read, :methreadline, :methtell, and :methseek),
-            or a string or os.PathLike object containing a file name
-        map_location: a function, torch.device, string or a dict specifying how to remap storage
-            locations
-        pickle_module: module used for unpickling metadata and objects (has to
-            match the pickle_module used to serialize file)
-        pickle_load_args: (Python 3 only) optional keyword arguments passed over to
-            pickle_module.load and pickle_module.Unpickler, e.g., errors=....
-
-    Example:
-
-    >>> better_torch_load('tensors.pt')
-    # Load all tensors onto the CPU
-    >>> better_torch_load('tensors.pt', map_location=torch.device('cpu'))
-    # Load all tensors onto the CPU, using a function
-    >>> better_torch_load('tensors.pt', map_location=lambda storage, loc: storage)
-    # Load all tensors onto GPU 1
-    >>> better_torch_load('tensors.pt', map_location=lambda storage, loc: storage.cuda(1))
-    # Map tensors from GPU 1 to GPU 0
-    >>> better_torch_load('tensors.pt', map_location={'cuda:1':'cuda:0'})
-    # Load tensor from io.BytesIO object
-    >>> with open('tensor.pt', 'rb') as f:
-            buffer = io.BytesIO(f.read())
-    >>> better_torch_load(buffer)
-    # Load a module with 'ascii' encoding for unpickling
-    >>> better_torch_load('module.pt', encoding='ascii')
-    """
-    # add yolov5 folder to system path
-    here = Path(__file__).parent.absolute()
-    yolov5_folder_dir = str(here)
-    sys.path.insert(0, yolov5_folder_dir)
-    # load torch model
-    torch_model = torch.load(f, map_location=map_location)
-    # remove yolov5 folder from system path
-    sys.path.remove(yolov5_folder_dir)
-    return torch_model
 
 
 class YOLOv5:
