+++ yolov5/models/experimental.py
@@ -6,9 +6,9 @@ from pathlib import Path
 import numpy as np
 import torch
 import torch.nn as nn
-from yolov5 import better_torch_load
 from yolov5.models.common import Conv, DWConv
 from yolov5.utils.google_utils import attempt_download
+from yolov5.utils.torch_utils import better_torch_load
 
 
 class CrossConv(nn.Module):
