+++ yolov5/train.py
@@ -22,7 +22,6 @@ from torch.utils.tensorboard import SummaryWriter
 from tqdm import tqdm
 
 import yolov5.test as test  # import test.py to get mAP after each epoch
-from yolov5 import better_torch_load
 from yolov5.models.experimental import attempt_load
 from yolov5.models.yolo import Model
 from yolov5.utils.autoanchor import check_anchors
@@ -38,7 +37,8 @@ from yolov5.utils.loss import ComputeLoss
 from yolov5.utils.neptuneai_logging.neptuneai_utils import NeptuneLogger
 from yolov5.utils.plots import (plot_evolution, plot_images, plot_labels,
                                 plot_results)
-from yolov5.utils.torch_utils import (ModelEMA, intersect_dicts, is_parallel,
+from yolov5.utils.torch_utils import (ModelEMA, better_torch_load,
+                                      intersect_dicts, is_parallel,
                                       select_device,
                                       torch_distributed_zero_first)
 from yolov5.utils.wandb_logging.wandb_utils import (WandbLogger,
