+++ yolov5/utils/aws/resume.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import torch
 import yaml
-from yolov5 import better_torch_load
+from yolov5.utils.torch_utils import better_torch_load
 
 sys.path.append('./')  # to run '$ python *.py' files in subdirectories
