+++ yolov5/utils/datasets.py
@@ -19,12 +19,12 @@ import torch.nn.functional as F
 from PIL import ExifTags, Image
 from torch.utils.data import Dataset
 from tqdm import tqdm
-from yolov5 import better_torch_load
 from yolov5.utils.general import (check_requirements, clean_str,
                                   resample_segments, segment2box,
                                   segments2boxes, xyn2xy, xywh2xyxy,
                                   xywhn2xyxy, xyxy2xywh)
-from yolov5.utils.torch_utils import torch_distributed_zero_first
+from yolov5.utils.torch_utils import (better_torch_load,
+                                      torch_distributed_zero_first)
 
 # Parameters
 help_url = 'https://github.com/ultralytics/yolov5/wiki/Train-Custom-Data'
