+++ yolov5/utils/general.py
@@ -20,10 +20,9 @@ import pkg_resources as pkg
 import torch
 import torchvision
 import yaml
-from yolov5 import better_torch_load
 from yolov5.utils.google_utils import gsutil_getsize
 from yolov5.utils.metrics import fitness
-from yolov5.utils.torch_utils import init_torch_seeds
+from yolov5.utils.torch_utils import better_torch_load, init_torch_seeds
 
 # Settings
 torch.set_printoptions(linewidth=320, precision=5, profile='long')
