+++ yolov5/utils/torch_utils.py
@@ -4,8 +4,10 @@ import datetime
 import logging
 import math
 import os
+import pickle
 import platform
 import subprocess
+import sys
 import time
 from contextlib import contextmanager
 from copy import deepcopy
@@ -301,3 +303,53 @@ class ModelEMA:
     def update_attr(self, model, include=(), exclude=('process_group', 'reducer')):
         # Update EMA attributes
         copy_attr(self.ema, model, include, exclude)
+
+
+def better_torch_load(f, map_location=None, pickle_module=pickle, **pickle_load_args):
+    """
+    Loads an object saved with torch.save from a file.
+
+    Same interface with https://pytorch.org/docs/stable/generated/torch.load.html
+    but better yolov5 model loading handling.
+
+    torch.hub handles it in the same way: https://github.com/pytorch/pytorch/blob/75024e228ca441290b6a1c2e564300ad507d7af6/torch/hub.py#L387
+    
+    Proper fix for: #22, #134, #353, #1155, #1389, #1680, #2531, #3071   
+    No need for such workarounds: #869, #1052, #2949
+
+    Args:
+        f: a file-like object (has to implement read, :methreadline, :methtell, and :methseek),
+            or a string or os.PathLike object containing a file name
+        map_location: a function, torch.device, string or a dict specifying how to remap storage
+            locations
+        pickle_module: module used for unpickling metadata and objects (has to
+            match the pickle_module used to serialize file)
+        pickle_load_args: (Python 3 only) optional keyword arguments passed over to
+            pickle_module.load and pickle_module.Unpickler, e.g., errors=....
+
+    Example:
+
+    >>> better_torch_load('tensors.pt')
+    # Load all tensors onto the CPU
+    >>> better_torch_load('tensors.pt', map_location=torch.device('cpu'))
+    # Load all tensors onto the CPU, using a function
+    >>> better_torch_load('tensors.pt', map_location=lambda storage, loc: storage)
+    # Load all tensors onto GPU 1
+    >>> better_torch_load('tensors.pt', map_location=lambda storage, loc: storage.cuda(1))
+    # Map tensors from GPU 1 to GPU 0
+    >>> better_torch_load('tensors.pt', map_location={'cuda:1':'cuda:0'})
+    # Load tensor from io.BytesIO object
+    >>> with open('tensor.pt', 'rb') as f:
+            buffer = io.BytesIO(f.read())
+    >>> better_torch_load(buffer)
+    # Load a module with 'ascii' encoding for unpickling
+    >>> better_torch_load('module.pt', encoding='ascii')
+    """
+    # add yolov5 folder to system path
+    yolov5_folder_dir = str(Path(__file__).parents[1].absolute())
+    sys.path.insert(0, yolov5_folder_dir)
+    # load torch model
+    torch_model = torch.load(f, map_location=map_location)
+    # remove yolov5 folder from system path
+    sys.path.remove(yolov5_folder_dir)
+    return torch_model
